+++ src/org/helioviewer/jhv/input/KeyInputEvent.java
@@ -6,6 +6,9 @@ public record KeyInputEvent(Key key, boolean shiftDown) {
         BACKSPACE,
         DELETE,
         N,
-        P
+        P,
+        X,
+        Y,
+        Z
     }
 }